Rename vacancyV2 to vacancy
