--- src/components/vacancy/vacancyActionTypes.js
@@ -1,4 +1,0 @@
-//Vacancies actions
-export const VACANCIES_LOADING = "VACANCIES_LOADING";
-export const GET_VACANCY = "GET_VACANCY";
-export const GET_VACANCIES = "GET_VACANCIES";
