+++ src/containers/vacancy/VacancyCreateOnCallContainer.js
@@ -11,7 +11,7 @@ import { listAssignment } from '../../store/actions/assignment'
 import { listRegion } from '../../store/actions/region'
 import { listRestriction } from '../../store/actions/restriction'
 import { createVacancy } from '../../store/actions/vacancy'
-import VacancyCreateOnCall from '../../components/vacancyV2/VacancyCreateOnCall'
+import VacancyCreateOnCall from '../../components/vacancy/VacancyCreateOnCall'
 import { selectCallById, selectProcessByCallId } from '../../store/selectors/call'
 import { convertObjetsToOptions } from '../../utils/selectorHelpers'
 import { selectAssignment } from '../../store/selectors/assignment'
