+++ src/containers/vacancy/VacancyDeleteContainer.js
@@ -7,7 +7,7 @@ import { clearErrors } from '../../store/actions/error'
 import { readCallV2 } from '../../store/actions/call'
 import { getProcess } from '../../store/actions/process'
 import { readVacancyV2, deleteVacancy } from '../../store/actions/vacancy'
-import VacancyDelete from '../../components/vacancyV2/VacancyDelete'
+import VacancyDelete from '../../components/vacancy/VacancyDelete'
 import { selectVacancyById } from '../../store/selectors/vacancy'
 import { selectCallByVacancyId } from '../../store/selectors/call'
 import { selectProcessByVacancyId } from '../../store/selectors/process'
